ai-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
